Create separate file for actionCreators
